rename and nits
