+++ requirements.txt
@@ -3,4 +3,4 @@ datasets
 tomli >= 1.1.0 ; python_version < "3.11"
 tensorboard
 sentencepiece
-tiktoken==0.4.0
+tiktoken
